refactor: drop circular dependency over evo/chainhelper.h
