+++ src/llmq/instantsend.cpp
@@ -13,7 +13,6 @@
 #include <chainparams.h>
 #include <consensus/validation.h>
 #include <dbwrapper.h>
-#include <evo/chainhelper.h>
 #include <index/txindex.h>
 #include <masternode/sync.h>
 #include <net_processing.h>
@@ -30,6 +29,9 @@
 
 using node::fImporting;
 using node::fReindex;
+
+// Forward declaration to break dependency over node/transaction.h
+std::pair<CTransactionRef, uint256> GetTransactionBlock(const uint256& hash, const CTxMemPool* const mempool);
 
 namespace llmq
 {
