+++ test/lint/lint-circular-dependencies.py
@@ -38,12 +38,7 @@ EXPECTED_CIRCULAR_DEPENDENCIES = (
     "dsnotificationinterface -> llmq/chainlocks -> node/blockstorage -> dsnotificationinterface",
     "evo/assetlocktx -> llmq/signing -> net_processing -> txmempool -> evo/assetlocktx",
     "evo/cbtx -> evo/simplifiedmns -> evo/cbtx",
-    "evo/chainhelper -> evo/specialtxman -> evo/deterministicmns -> evo/chainhelper",
     "evo/chainhelper -> evo/specialtxman -> validation -> evo/chainhelper",
-    "evo/chainhelper -> llmq/chainlocks -> evo/chainhelper",
-    "evo/chainhelper -> llmq/instantsend -> evo/chainhelper",
-    "evo/chainhelper -> masternode/payments -> governance/classes -> governance/object -> evo/chainhelper",
-    "evo/chainhelper -> node/transaction -> node/context -> evo/chainhelper",
     "evo/deterministicmns -> llmq/commitment -> evo/deterministicmns",
     "evo/deterministicmns -> llmq/commitment -> validation -> evo/deterministicmns",
     "evo/deterministicmns -> llmq/commitment -> validation -> txmempool -> evo/deterministicmns",
